# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个使用Go、Gin、Gorm和PostgreSQL构建的Web后端项目，支持通过Phantom钱包连接用户，并实现了JWT认证。",
        "title": "Go Web Backend API",
        "termsOfService": "http://your-terms-of-service-url.com",
        "contact": {
            "name": "API Support",
            "url": "http://www.yourwebsite.com/support",
            "email": "support@yourwebsite.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9100",
    "paths": {
...
        "/api/connect_wallet": {
            "post": {
                "description": "用户通过Phantom钱包连接到系统，创建或更新用户信息，并返回JWT令牌。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "连接Phantom钱包",
                "parameters": [
                    {
                        "description": "用户连接Phantom钱包请求体",
                        "name": "connect_wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ConnectWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户已连接",
                        "schema": {
                            "$ref": "#/definitions/handlers.ConnectWalletResponse"
                        }
                    },
                    "201": {
                        "description": "用户已创建",
                        "schema": {
                            "$ref": "#/definitions/handlers.ConnectWalletResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前认证用户的详细资料。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户资料",
                "responses": {
                    "200": {
                        "description": "用户资料",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetProfileResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "string",
            "enum": [
                "INVALID_REQUEST",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "INTERNAL_ERROR",
                "DATABASE_ERROR",
                "VALIDATION_ERROR",
                "TOKEN_GENERATION_ERROR",
                "TOKEN_VERIFICATION_ERROR"
            ],
            "x-enum-varnames": [
                "ErrInvalidRequest",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrInternal",
                "ErrDatabase",
                "ErrValidation",
                "ErrTokenGeneration",
                "ErrTokenVerification"
            ]
        },
...
        "handlers.ConnectWalletRequest": {
            "type": "object",
            "required": [
                "message",
                "signature",
                "wallet_address"
            ],
            "properties": {
                "message": {
                    "description": "Message 签名消息（必填，用于验证 nonce）",
                    "type": "string"
                },
                "signature": {
                    "description": "Signature 用户签名（必填）",
                    "type": "string"
                },
                "username": {
                    "description": "Username 用户名（可选）",
                    "type": "string",
                    "maxLength": 50
                },
                "wallet_address": {
                    "description": "WalletAddress 即公钥（必填）",
                    "type": "string"
                }
            }
        },
        "handlers.ConnectWalletResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message 操作消息",
                    "type": "string"
                },
                "token": {
                    "description": "Token JWT令牌",
                    "type": "string"
                },
                "user": {
                    "description": "User 用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "handlers.GetProfileResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "description": "User 用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}